+++ src/extensionSupport.test.ts
@@ -1,12 +1,15 @@
 import assert from "node:assert/strict";
 import {
   COMPLETION_TRIGGER_CHARACTERS,
+  DIRECTIVE_VALUE_TRIGGER_CHARACTERS,
+  directiveValueContextAt,
   isTxtJetPath,
   selectedTargetLanguageId,
   shouldOfferMarkerCompletions
 } from "./extensionSupport";
 
 assert.deepEqual(COMPLETION_TRIGGER_CHARACTERS, ["<"]);
+assert.deepEqual(DIRECTIVE_VALUE_TRIGGER_CHARACTERS, ["\"", "'", "/", "\\", "."]);
 
 assert.equal(shouldOfferMarkerCompletions("<"), true);
 assert.equal(shouldOfferMarkerCompletions("prefix <"), true);
@@ -28,5 +31,34 @@ assert.equal(isTxtJetPath("vscode-remote://ssh-remote+host/workspace/example.txt
 assert.equal(selectedTargetLanguageId("txtjet-html", "txtjet-java"), "txtjet-html");
 assert.equal(selectedTargetLanguageId("txtjet", "txtjet-python"), "txtjet");
 assert.equal(selectedTargetLanguageId("plaintext", "txtjet-python"), "txtjet-python");
+
+const includeDirective = "<%@ include file=\"partials/head\" %>";
+assert.deepEqual(directiveValueContextAt(includeDirective, includeDirective.indexOf("head") + 4), {
+  directiveName: "include",
+  attributeName: "file",
+  value: "partials/head",
+  valueRange: {
+    start: includeDirective.indexOf("partials/head"),
+    end: includeDirective.indexOf("partials/head") + "partials/head".length
+  },
+  quote: "\"",
+  prefix: "partials/head"
+});
+
+const skeletonDirective = "<%@ jet class=\"Demo\" skeleton='templates/base' %>";
+assert.deepEqual(directiveValueContextAt(skeletonDirective, skeletonDirective.indexOf("base") + 2), {
+  directiveName: "jet",
+  attributeName: "skeleton",
+  value: "templates/base",
+  valueRange: {
+    start: skeletonDirective.indexOf("templates/base"),
+    end: skeletonDirective.indexOf("templates/base") + "templates/base".length
+  },
+  quote: "'",
+  prefix: "templates/ba"
+});
+
+assert.equal(directiveValueContextAt("<%@ include file=\"done\" %>", "<%@ include file=\"done\" %>".length), undefined);
+assert.equal(directiveValueContextAt("<%= value %>", 4), undefined);
 
 console.log("extension support tests ok");